feat(prelude): add StackVec, replace smallvec dependency refactor(prelude): model Board as a street enum feat(prelude): add HoleCardRule and runtime starting-hand tables refactor(prelude): reorganize suit-isomorphism into GTO/EV forms feat(prelude): add Waugh HandIndexer refactor(prelude): flatten and regroup module tree fix(just): unbreak coverage recipe on lcov 2.5

+++ openpql-core/src/functions/equity/fractional_river_equity.rs
@@ -18,7 +18,7 @@ pub fn fractional_river_equity(
     hero_idx: PQLPlayerCount,
 ) -> PQLFraction {
     debug_assert!(
-        board.len() == PQLBoard::N_RIVER,
+        board.len() == PQLBoard::N_RIVER as usize,
         "board must have {} cards, got {}",
         PQLBoard::N_RIVER,
         board.len(),
